fast-forwards

+++ submodule.py
@@ -5,8 +5,8 @@ from typing import Optional
 
 class Foo(BaseModel):
     count: int
-    size: Optional[float] = None
-    identifier: float = 0.0
+    size: Optional[int] = None
+    identifier: int = 0.0
 
 
 
@@ -16,7 +16,7 @@ class Foo(BaseModel):
 class Bar(BaseModel):
     apple: str = 'x'
     banana: str = 'y'
-    identifier: float = 0.0
+    identifier: int = 0.0
 
 
 
@@ -25,7 +25,7 @@ class Bar(BaseModel):
 class Spam(BaseModel):
     bars: list[Bar]
     name: str  = "Spam"
-    identifier: float = 0.0
+    identifier: int = 0.0
 
 
 m = Spam(foo={'count': 4}, bars=[{'apple': 'x1'}, {'apple': 'x2'}])
